tessarest/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
Pillow==9.3.0
requests==2.27.1
